* created separate files for different model configurations * moved loss functions into model architecture class
