+++ src/models/architecture/cnn.py
@@ -1,0 +1,89 @@
+import numpy as np
+import tensorflow as tf
+import tensorflow_probability as tfp
+from tensorflow.python import keras
+from tensorflow.python.keras import layers
+
+
+class CNN(keras.Model):
+    def __init__(self, action_size, ip_shape=(84, 84, 3)):
+        super(CNN, self).__init__()
+        self.action_size = action_size
+        self.ip_shape = ip_shape
+
+        # common network with shared parameters
+        # (20, 20, 16)
+        self.conv1 = layers.Conv2D(filters=16, kernel_size=(8, 8), strides=(
+            4, 4), padding='same', activation='relu', data_format='channels_last', input_shape=self.ip_shape)
+        # (9, 9, 32)
+        self.conv2 = layers.Conv2D(filters=32, kernel_size=(4, 4), strides=(
+            2, 2), padding='same', activation='relu', data_format='channels_last')
+        # (9 * 9 * 32)
+        self.flatten = layers.Flatten()
+        self.dense1 = layers.Dense(units=256, activation='relu')
+
+        # policy output layer (Actor)
+        self.policy_logits = layers.Dense(
+            self.action_size, name='policy_logits')
+
+        # value output layer (Critic)
+        self.values = layers.Dense(units=1, name='value')
+
+    def call(self, inputs):
+        x = self.conv1(inputs)
+        x = self.conv2(x)
+        x = self.flatten(x)
+        x = self.dense1(x)
+
+        logits = self.policy_logits(x)
+        values = self.values(x)
+
+        return logits, values
+
+    def compute_loss(self, done, new_state, memory, gamma=0.99):
+        beta_regularizer = 0.01
+
+        if done:  # game has terminated
+            reward_sum = 0.
+        else:
+            reward_sum = self.call(tf.convert_to_tensor(
+                np.expand_dims(new_state, axis=0), dtype=tf.float32))[-1].numpy()[0]
+
+        # Get discounted rewards
+        # TODO: try to normalize the discounted rewards
+        discounted_rewards = []
+        for reward in memory.rewards[::-1]:  # reverse buffer r
+            reward_sum = reward + gamma * reward_sum
+            discounted_rewards.append(reward_sum)
+        discounted_rewards.reverse()
+
+        logits, values = self.call(tf.convert_to_tensor(
+            np.stack(memory.states), dtype=tf.float32))
+
+        # get our advantages
+        q_value_estimate = tf.convert_to_tensor(
+            np.array(discounted_rewards)[:, None], dtype=tf.float32)
+        advantage = q_value_estimate - values
+
+        # value loss
+        value_loss = advantage ** 2
+
+        # policy loss
+        actions_one_hot = tf.one_hot(
+            memory.actions, self.action_size, dtype=tf.float32)
+
+        policy = tf.nn.softmax(logits)
+        entropy = tf.nn.softmax_cross_entropy_with_logits(
+            labels=policy, logits=logits)
+        # entropy = tf.reduce_sum(policy * tf.math.log(policy + 1e-20), axis=1)
+
+        policy_loss = tf.nn.sparse_softmax_cross_entropy_with_logits(
+            labels=memory.actions, logits=logits)
+        # policy_loss = tf.compat.v1.nn.softmax_cross_entropy_with_logits_v2(
+        #     labels=actions_one_hot, logits=logits)
+        policy_loss *= tf.stop_gradient(advantage)
+        policy_loss -= beta_regularizer * entropy
+
+        total_loss = tf.reduce_mean((0.5 * value_loss + policy_loss))
+
+        return total_loss
